+++ include/minirt.h
@@ -6,7 +6,7 @@
 /*   By: akdovlet <akdovlet@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/16 16:12:46 by akdovlet          #+#    #+#             */
-/*   Updated: 2025/01/14 18:35:24 by akdovlet         ###   ########.fr       */
+/*   Updated: 2025/01/15 14:09:29 by akdovlet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,9 +33,9 @@
 # define HEIGHT 1080
 
 t_ray		ray_new(t_tuple origin, t_tuple direction);
-t_tuple		position(t_ray ray, float factor);
+t_tuple		position(t_ray ray, double factor);
 void		render(t_camera cam, t_world world, t_img *img, t_mlx *mlx);
-t_ray		ray_for_pixel(t_camera cam, float x, float y);
+t_ray		ray_for_pixel(t_camera cam, double x, double y);
 bool		is_shadowed(t_world world, t_tuple point);
 t_tuple		ring_at(t_pattern ring, t_tuple point);
 t_tuple		gradient_at(t_pattern pattern, t_tuple point);
