+++ src/camera/camera.c
@@ -6,19 +6,19 @@
 /*   By: akdovlet <akdovlet@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/05 18:17:18 by akdovlet          #+#    #+#             */
-/*   Updated: 2025/01/14 18:22:38 by akdovlet         ###   ########.fr       */
+/*   Updated: 2025/01/15 18:32:40 by akdovlet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt.h"
 
-t_camera	camera_new(float hsize, float vsize, float fov)
+t_camera	camera_new(double hsize, double vsize, double fov)
 {
 	t_camera	cam;
-	float		aspect_ratio;
-	float		half_view;
+	double		aspect_ratio;
+	double		half_view;
 
-	half_view = tanf(fov / 2.0f);
+	half_view = tan(fov / 2.0);
 	aspect_ratio = hsize / vsize;
 	if (aspect_ratio >= 1)
 	{
