+++ src/matrix/transform.c
@@ -6,13 +6,13 @@
 /*   By: akdovlet <akdovlet@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/26 10:22:22 by akdovlet          #+#    #+#             */
-/*   Updated: 2025/01/04 15:12:33 by akdovlet         ###   ########.fr       */
+/*   Updated: 2025/01/15 14:09:10 by akdovlet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt.h"
 
-t_matrix	scale(float x, float y, float z)
+t_matrix	scale(double x, double y, double z)
 {
 	return ((t_matrix){{
 			{x, 0, 0, 0}, \
@@ -22,7 +22,7 @@ t_matrix	scale(float x, float y, float z)
 	}});
 }
 
-t_matrix	translate(float x, float y, float z)
+t_matrix	translate(double x, double y, double z)
 {
 	return ((t_matrix){{
 			{1, 0, 0, x}, \
