+++ src/mlx/mlx_pixel_put.c
@@ -6,13 +6,13 @@
 /*   By: akdovlet <akdovlet@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/18 16:41:45 by akdovlet          #+#    #+#             */
-/*   Updated: 2025/01/10 20:15:25 by akdovlet         ###   ########.fr       */
+/*   Updated: 2025/01/15 14:09:10 by akdovlet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt.h"
 
-int	is_within_bounds(float x, float y)
+int	is_within_bounds(double x, double y)
 {
 	if (x < WIDTH && y < HEIGHT && x > 0 && y > 0)
 		return (1);
