+++ src/pattern/gradient.c
@@ -6,7 +6,7 @@
 /*   By: akdovlet <akdovlet@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/10 20:19:49 by akdovlet          #+#    #+#             */
-/*   Updated: 2025/01/13 10:01:13 by akdovlet         ###   ########.fr       */
+/*   Updated: 2025/01/15 18:33:20 by akdovlet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,11 +15,11 @@
 t_tuple	gradient_at(t_pattern pattern, t_tuple point)
 {
 	t_tuple	distance;
-	float	fraction;
+	double	fraction;
 	t_tuple	color;
 	
-	fraction = 2.0f * point.x - floorf(2.0f * point.x);
-	if(fabsf(point.x - floorf(point.x)) < 0.5f)
+	fraction = 2.0 * point.x - floor(2.0 * point.x);
+	if(fabs(point.x - floor(point.x)) < 0.5)
 	{
 		distance = tuple_substract(pattern.color2, pattern.color1);
 		color = tuple_add(pattern.color1, tuple_multiply(distance, fraction));
