+++ src/tuple/color_new.c
@@ -6,13 +6,13 @@
 /*   By: akdovlet <akdovlet@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/17 18:15:41 by akdovlet          #+#    #+#             */
-/*   Updated: 2024/12/19 15:23:18 by akdovlet         ###   ########.fr       */
+/*   Updated: 2025/01/15 14:09:10 by akdovlet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt.h"
 
-t_tuple	color_new(float red, float green, float blue)
+t_tuple	color_new(double red, double green, double blue)
 {
 	t_tuple	new;
