+++ src/tuple/tuple_new.c
@@ -6,13 +6,13 @@
 /*   By: akdovlet <akdovlet@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/16 16:15:30 by akdovlet          #+#    #+#             */
-/*   Updated: 2024/12/19 15:24:17 by akdovlet         ###   ########.fr       */
+/*   Updated: 2025/01/15 14:09:10 by akdovlet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt.h"
 
-t_tuple	tuple_new(float x, float y, float z, float w)
+t_tuple	tuple_new(double x, double y, double z, double w)
 {
 	t_tuple	new;
 
@@ -23,12 +23,12 @@ t_tuple	tuple_new(float x, float y, float z, float w)
 	return (new);
 }
 
-t_tuple	vector_new(float x, float y, float z)
+t_tuple	vector_new(double x, double y, double z)
 {
 	return (tuple_new(x, y, z, 0));
 }
 
-t_tuple	point_new(float x, float y, float z)
+t_tuple	point_new(double x, double y, double z)
 {
 	return (tuple_new(x, y, z, 1));
 }
