+++ src/tuple/tuple_scalar.c
@@ -6,13 +6,13 @@
 /*   By: akdovlet <akdovlet@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/12/16 18:15:14 by akdovlet          #+#    #+#             */
-/*   Updated: 2024/12/19 15:23:33 by akdovlet         ###   ########.fr       */
+/*   Updated: 2025/01/15 14:09:10 by akdovlet         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minirt.h"
 
-t_tuple	tuple_multiply(t_tuple a, float factor)
+t_tuple	tuple_multiply(t_tuple a, double factor)
 {
 	t_tuple	new;
 
@@ -23,7 +23,7 @@ t_tuple	tuple_multiply(t_tuple a, float factor)
 	return (new);
 }
 
-t_tuple	tuple_divide(t_tuple a, float divisor)
+t_tuple	tuple_divide(t_tuple a, double divisor)
 {
 	t_tuple	new;
